pyproject.toml
[tool.poetry]
name = "finances-dw"
version = "0.1.0"
description = ""
authors = ["Paco <pacodeoliveirasaavedra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
plotly = "^5.24.0"
streamlit = "^1.38.0"
yfinance = "^0.2.44"
sqlalchemy = "^2.0.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
